Realises that the current match function is really a case statement.
Cases, withs the like are higher level fns that are implemented by
calling something like bindings and passing the result of that to a fn.

This also sketches out an idea for patterns as structs - allowing us to
define and extend custom patterns and implementations for those patterns
for any given data type.

Time to do two things:

1. Clean up the idea of matching from case statements.
2. Explore the matches as structs idea. This is tricky to do without
   drastically changing the library. Which is fine to do and all, but I
   want to be able to see both versions side by side to compare. Guess
   we can branch...